+++ packages/styles/src/packages/select/index.ts
@@ -1,0 +1,5 @@
+import * as styles from "./select.style.css";
+
+export const select = {
+  style: styles,
+};